Updates to community member policy and views
Slight overhaul to the policy that determines if a user has access to a community. The policy now adds the highest level role the user has for the community as a local for use in controllers and views, makes for simpler and consistent conditionals. Adds logic to views for admins/superusers. Updates some styling as well.